Create a module called affine.py
